refactor(Header): Update UI to match Figma design
