+++ public/manifest.json
@@ -24,5 +24,6 @@
   },
   "options_page": "options.html",
   "permissions": ["activeTab", "storage"],
-  "content_security_policy": "script-src 'self' 'unsafe-eval'; object-src 'self'"
+  "content_security_policy": "script-src 'self' 'unsafe-eval'; object-src 'self'",
+  "web_accessible_resources": ["*.ttf" ]
 }
