+++ public/options.html
@@ -5,7 +5,7 @@
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
-      content="extension created using create-react-extension"
+      content="Options for V Rising Server Scanner"
     />
     <link rel="apple-touch-icon" href="logo192.png" />
     <!--
@@ -22,7 +22,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <title>V Rising Server Scanner Options</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
