+++ public/popup.html
@@ -5,7 +5,7 @@
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
-      content="extension created using create-react-extension"
+      content="Save and track dedicated V Rising Servers."
     />
     <link rel="apple-touch-icon" href="logo192.png" />
     <!--
@@ -22,7 +22,10 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Paytone+One&display=swap" rel="stylesheet">
+    <title>V Rising Server Scanner</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
