+++ src/layouts/PopupLayout/Header/constant.js
@@ -1,5 +1,5 @@
-export const HEADER_TITLE = 'V Rising Server Scanner';
+export const HEADER_TITLE = 'RISING';
 
-export const HEADER_DESCRIPTION = 'Search and save dedicated V Rising servers for easy access to server information';
+export const HEADER_DESCRIPTION = 'SERVER SCANNER';
 
 export const HEADER_HEIGHT = 105;
